@@ -5,32 +5,37 @@
   "author": "Matthew Hinton <mhinton@gmail.com>",
   "private": true,
   "scripts": {
-    "dev": "PORT=4000 node server.js",
-    "build": "nuxt build",
-    "start": "nuxt start",
+    "Xdev-express": "PORT=4000 node server.js",
+    "old-dev": "PORT=4000 nuxt dev",
+    "dev": "PORT=4000 backpack dev",
+    "build": "nuxt build && backpack build",
+    "start": "NODE_ENV=production node build/main.js",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
     "precommit": "npm run lint"
   },
   "dependencies": {
     "@nuxtjs/bootstrap-vue": "^2.0.2",
-    "axios": "^0.16.2",
+    "axios": "^0.17.0",
     "body-parser": "^1.18.2",
     "debug": "^3.1.0",
-    "express": "^4.16.1",
+    "express": "^4.16.2",
     "express-session": "^1.15.6",
     "nuxt": "^1.0.0-rc11",
-    "startbootstrap-bare": "^4.0.0-beta",
-    "vuex": "^2.4.1"
+    "vue": "^2.5.2",
+    "vue-server-renderer": "^2.5.2",
+    "vue-template-compiler": "^2.5.2",
+    "vuex": "^3.0.0"
   },
   "devDependencies": {
     "babel-eslint": "^8.0.1",
-    "eslint": "^4.8.0",
-    "eslint-config-airbnb-base": "^12.0.2",
+    "backpack-core": "^0.4.3",
+    "eslint": "^4.9.0",
+    "eslint-config-airbnb-base": "^12.1.0",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-html": "^3.2.2",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.5.0"
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0"
   }
 }
