@@ -32,32 +32,34 @@ export default {
     return {
       formError: null,
       formUsername: '',
-      formPassword: ''
-    }
+      formPassword: '',
+    };
   },
+
   methods: {
     async login() {
       try {
         await this.$store.dispatch('login', {
           username: this.formUsername,
-          password: this.formPassword
-        })
-        this.formUsername = ''
-        this.formPassword = ''
-        this.formError = null
+          password: this.formPassword,
+        });
+        this.formUsername = '';
+        this.formPassword = '';
+        this.formError = null;
       } catch (e) {
-        this.formError = e.message
+        this.formError = e.message;
       }
     },
+
     async logout() {
       try {
-        await this.$store.dispatch('logout')
+        await this.$store.dispatch('logout');
       } catch (e) {
-        this.formError = e.message
+        this.formError = e.message;
       }
-    }
-  }
-}
+    },
+  },
+};
 </script>
 
 <style>
